+++ protected/humhub/commands/SettingsController.php
@@ -8,15 +8,12 @@
 
 namespace humhub\commands;
 
-use humhub\components\Module;
 use humhub\components\SettingsManager;
-use humhub\libs\DynamicConfig;
 use humhub\models\Setting;
-use Yii;
 use yii\console\Controller;
 use yii\console\ExitCode;
-use yii\helpers\Console;
 use yii\console\widgets\Table;
+use yii\helpers\Console;
 
 /**
  * SettingsController provides CLI access to database settings
@@ -88,8 +85,6 @@ class SettingsController extends Controller
         $settingsManager = new SettingsManager(['moduleId' => $moduleId]);
         $settingsManager->set($name, $value);
 
-        $this->handleDynamicConfig($moduleId, $name);
-
         $this->stdout("\n*** Successfully set setting\n\n", Console::FG_GREEN);
         $this->stdout("Name:\t\t" . $name . "\n");
         $this->stdout("Module ID:\t" . $moduleId . "\n\n");
@@ -112,7 +107,6 @@ class SettingsController extends Controller
 
         $settingsManager = new SettingsManager(['moduleId' => $moduleId]);
         $settingsManager->delete($name);
-        $this->handleDynamicConfig($moduleId, $name);
 
         $this->stdout("\n*** Successfully deleted setting\n\n", Console::FG_GREEN);
         $this->stdout("Name:\t\t" . $name . "\n");
@@ -120,20 +114,4 @@ class SettingsController extends Controller
 
         return ExitCode::OK;
     }
-
-    /**
-     * Handles dynamic config rewrite if required
-     *
-     * @param $moduleId
-     * @param $name
-     */
-    private function handleDynamicConfig($moduleId, $name)
-    {
-        if (DynamicConfig::needRewrite($moduleId, $name)) {
-            // Force reload
-            Yii::$app->settings->init();
-            DynamicConfig::rewrite();
-        }
-    }
-
 }