+++ protected/humhub/components/Application.php
@@ -8,9 +8,9 @@
 
 namespace humhub\components;
 
+use Exception;
 use humhub\interfaces\ApplicationInterface;
 use Yii;
-use Exception;
 
 /**
  * Description of Application
@@ -69,7 +69,7 @@ class Application extends \yii\web\Application implements ApplicationInterface
         /**
          * Check if it's already installed - if not force controller module
          */
-        if (!$this->params['installed'] && $this->controller->module != null && $this->controller->module->id != 'installer') {
+        if (!$this->installationState->hasState(InstallationState::STATE_INSTALLED) && $this->controller && $this->controller->module != null && $this->controller->module->id != 'installer') {
             $this->controller->redirect(['/installer/index']);
             return false;
         }