+++ protected/humhub/modules/admin/models/forms/BasicSettingsForm.php
@@ -2,7 +2,6 @@
 
 namespace humhub\modules\admin\models\forms;
 
-use humhub\libs\DynamicConfig;
 use humhub\modules\ui\icon\widgets\Icon;
 use Yii;
 use yii\base\Model;
@@ -111,8 +110,6 @@ class BasicSettingsForm extends Model
         Yii::$app->getModule('dashboard')->settings->set('showProfilePostForm', $this->dashboardShowProfilePostForm);
         Yii::$app->getModule('tour')->settings->set('enable', $this->tour);
         Yii::$app->getModule('friendship')->settings->set('enable', $this->enableFriendshipModule);
-
-        DynamicConfig::rewrite();
 
         return true;
     }